Use filters/photflam from settings, load data in subprocesses
To be sure that any band images and filter can be provided, these must
now be specified in the settings file.

Also most data (cube, hr images) are now loaded in the Deblend class,
which removes the need to prepare inputs before. And avoids the cost of
pickling the data.