Fix issue with missing sources in rotated segmap
When the segmap needs to be rotated there were sometimes an issue with
missing sources in the deblend step. During grouping the full segmap is
rotated, but during deblend it is cropped first. And the cropping was
too restrictive when the cutout is not square.

Also fix a potential issue when the HR image is not aligned with North.